Refactor: 의존성 순환 해결
AuthService와 TokenProvider의 의존성 순환을 해결하기 위해 TokenAuthenticator 추가
