+++ api/src/main/java/com/zerobase/commerce/api/exception/ErrorResponse.java
@@ -35,9 +35,9 @@ public class ErrorResponse {
     @Getter
     @AllArgsConstructor
     static class ErrorResponseDto {
+        private final LocalDateTime timeStamp;
         private String errorCode;
         private String errorMessage;
-        private final LocalDateTime timeStamp;
 
         public static ErrorResponseDtoBuilder builder() {
             return new ErrorResponseDtoBuilder();
