+++ api/src/test/java/com/zerobase/commerce/api/user/controller/UserControllerTest.java
@@ -1,9 +1,12 @@
 package com.zerobase.commerce.api.user.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.zerobase.commerce.api.security.TokenAuthenticator;
 import com.zerobase.commerce.api.security.TokenProvider;
-import com.zerobase.commerce.api.user.dto.*;
+import com.zerobase.commerce.api.user.dto.UpdateUserInfo;
+import com.zerobase.commerce.api.user.dto.UserDto;
 import com.zerobase.commerce.api.user.service.UserService;
+import com.zerobase.commerce.database.constant.AuthorityStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,8 +15,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.Arrays;
-import java.util.stream.Collectors;
+import java.util.HashSet;
+import java.util.Set;
 
 import static com.zerobase.commerce.database.constant.AuthorityStatus.ROLE_MEMBER;
 import static org.mockito.ArgumentMatchers.any;
@@ -38,57 +41,20 @@ class UserControllerTest {
     @MockBean
     private TokenProvider tokenProvider;
 
-    @Test
-    @DisplayName("회원가입 성공")
-    void signUpSuccess() throws Exception {
-        var request = SignUpDto.Request.builder()
-                .id("root1234")
-                .password("qwer1234")
-                .build();
-
-        given(userService.signUp(any())).willReturn(
-                SignUpDto.Response.builder()
-                        .id("root1234")
-                        .password("qwer1234")
-                        .roles(Arrays.stream(new String[]{"ROLE_MEMBER"}).collect(Collectors.toSet()))
-                        .build());
-
-        var result = mockMvc.perform(post("/auth/sign-up")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)));
-
-        result.andExpect(status().isOk())
-                .andExpect(jsonPath("$.id").value("root1234"))
-                .andExpect(jsonPath("$.password").value("qwer1234"))
-                .andExpect(jsonPath("$.roles[0]").value(ROLE_MEMBER.name()))
-                .andDo(print());
-    }
-
-    @Test
-    @DisplayName("로그인 성공")
-    void signInSuccess() throws Exception {
-        var request = SignInDto.Request.builder()
-                .id("root1234")
-                .password("qwer1234")
-                .build();
-
-        given(userService.signIn(any())).willReturn("token");
-
-        var result = mockMvc.perform(post("/auth/sign-in")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)));
-
-        result.andExpect(status().isOk());
-    }
+    @MockBean
+    private TokenAuthenticator tokenAuthenticator;
 
     @Test
     @DisplayName("유저 정보 가져오기 성공")
     void getUserInfoSuccess() throws Exception {
+        Set<AuthorityStatus> roles = new HashSet<>();
+        roles.add(ROLE_MEMBER);
+
         given(userService.getUserInfo(any(), anyString())).willReturn(
                 UserDto.builder()
                         .id("root1234")
                         .password("qwer1234")
-                        .roles(Arrays.stream(new String[]{"ROLE_MEMBER"}).collect(Collectors.toSet()))
+                        .roles(roles)
                         .build()
         );
 
@@ -106,11 +72,14 @@ class UserControllerTest {
     @Test
     @DisplayName("유저 정보 수정 성공")
     void updateUserInfoSuccess() throws Exception {
+        Set<AuthorityStatus> roles = new HashSet<>();
+        roles.add(ROLE_MEMBER);
+
         given(userService.updateUserInfo(any(), any())).willReturn(
                 UserDto.builder()
                         .id("root1234")
                         .password("zxcv1234")
-                        .roles(Arrays.stream(new String[]{"ROLE_MEMBER"}).collect(Collectors.toSet()))
+                        .roles(roles)
                         .build()
         );
