+++ api/src/test/java/com/zerobase/commerce/api/user/service/AuthServiceTest.java
@@ -2,9 +2,9 @@ package com.zerobase.commerce.api.user.service;
 
 import com.zerobase.commerce.api.exception.CustomException;
 import com.zerobase.commerce.api.exception.ErrorCode;
-import com.zerobase.commerce.api.security.TokenAuthenticator;
-import com.zerobase.commerce.api.user.dto.UpdateUserInfo;
-import com.zerobase.commerce.api.user.dto.UserDto;
+import com.zerobase.commerce.api.security.TokenProvider;
+import com.zerobase.commerce.api.user.dto.SignInDto;
+import com.zerobase.commerce.api.user.dto.SignUpDto;
 import com.zerobase.commerce.database.constant.AuthorityStatus;
 import com.zerobase.commerce.database.domain.User;
 import com.zerobase.commerce.database.repository.UserRepository;
@@ -23,8 +23,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import static com.zerobase.commerce.database.constant.AuthorityStatus.ROLE_MEMBER;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
@@ -33,19 +32,19 @@ import static org.mockito.Mockito.verify;
 
 @Transactional
 @ExtendWith(MockitoExtension.class)
-class UserServiceTest {
+class AuthServiceTest {
     @Mock
     UserRepository userRepository;
 
     @Mock
-    TokenAuthenticator tokenAuthenticator;
+    TokenProvider tokenProvider;
 
     @InjectMocks
-    UserService userService;
+    AuthService authService;
 
     @Test
-    @DisplayName("유저 정보 가져오기 성공")
-    void getUserInfoSuccess() {
+    @DisplayName("회원가입 성공")
+    void signUpSuccess() {
         // given
         Set<AuthorityStatus> roles = new HashSet<>();
         roles.add(ROLE_MEMBER);
@@ -58,20 +57,32 @@ class UserServiceTest {
                 .updatedAt(LocalDateTime.now())
                 .build();
 
-        given(tokenAuthenticator.resolveTokenFromHeader(any())).willReturn("asdf1234");
-        given(userRepository.findById(anyString())).willReturn(Optional.of(user));
+        var request = SignUpDto.Request.builder()
+                .id("asdf1234")
+                .password("qwer1234")
+                .build();
+
+        given(userRepository.existsById(anyString())).willReturn(false);
+        given(userRepository.save(any())).willReturn(user);
+
+        ArgumentCaptor<User> captor = ArgumentCaptor.forClass(User.class);
 
         // when
-        UserDto userDto = userService.getUserInfo(any(), "qwer1234");
+        SignUpDto.Response response = authService.signUp(request);
 
         // then
-        assertEquals(userDto.getId(), user.getId());
-        assertEquals(userDto.getPassword(), user.getPassword());
+        verify(userRepository, times(1)).save(captor.capture());
+
+        assertEquals(response.getId(), user.getId());
+        assertEquals(response.getPassword(), user.getPassword());
+
+        assertEquals(captor.getValue().getId(), user.getId());
+        assertEquals(captor.getValue().getPassword(), user.getPassword());
     }
 
     @Test
-    @DisplayName("유저 정보 가져오기 실패 - 아이디를 찾을 수 없음")
-    void getUserInfoFailedIdNotFound() {
+    @DisplayName("회원가입 실패 - 중복 아이디 입력")
+    void signUpFailedDuplicateId() {
         // given
         Set<AuthorityStatus> roles = new HashSet<>();
         roles.add(ROLE_MEMBER);
@@ -84,22 +95,88 @@ class UserServiceTest {
                 .updatedAt(LocalDateTime.now())
                 .build();
 
-        given(tokenAuthenticator.resolveTokenFromHeader(any())).willReturn("asdf1233");
+
+        given(userRepository.existsById(anyString())).willReturn(true);
+
+        // when
+        CustomException exception = assertThrows(
+                CustomException.class,
+                () -> authService.signUp(
+                        SignUpDto.Request.builder()
+                                .id(user.getId())
+                                .password(user.getPassword())
+                                .build()
+                )
+        );
+
+        // then
+        verify(userRepository, times(0)).save(any());
+        assertEquals(exception.getErrorCode(), ErrorCode.USER_ID_DUPLICATED);
+    }
+
+    @Test
+    @DisplayName("로그인 성공")
+    void signInSuccess() {
+        // given
+        Set<AuthorityStatus> roles = new HashSet<>();
+        roles.add(ROLE_MEMBER);
+
+        User user = User.builder()
+                .id("asdf1234")
+                .password("qwer1234")
+                .roles(roles)
+                .registeredAt(LocalDateTime.now())
+                .updatedAt(LocalDateTime.now())
+                .build();
+
+        given(userRepository.findById(anyString())).willReturn(Optional.of(user));
+        given(tokenProvider.generateToken(anyString(), any())).willReturn("token");
+
+        // when
+        String response = authService.signIn(SignInDto.Request.builder()
+                .id("asdf1234")
+                .password("qwer1234")
+                .build()
+        );
+
+        // then
+        assertNotNull(response);
+    }
+
+    @Test
+    @DisplayName("로그인 실패 - 아이디를 찾을 수 없음")
+    void signInFailedIdNotFound() {
+        // given
+        Set<AuthorityStatus> roles = new HashSet<>();
+        roles.add(ROLE_MEMBER);
+
+        User user = User.builder()
+                .id("asdf1234")
+                .password("qwer1234")
+                .roles(roles)
+                .registeredAt(LocalDateTime.now())
+                .updatedAt(LocalDateTime.now())
+                .build();
+
         given(userRepository.findById(anyString())).willReturn(Optional.empty());
 
         // when
         CustomException exception = assertThrows(
                 CustomException.class,
-                () -> userService.getUserInfo(any(), "qwer1234")
+                () -> authService.signIn(
+                        SignInDto.Request.builder()
+                                .id(user.getId())
+                                .password(user.getPassword())
+                                .build())
         );
 
         // then
-        assertEquals(exception.getErrorCode(), ErrorCode.INVALID_USER_ID);
+        assertEquals(exception.getErrorCode(), ErrorCode.SIGN_IN_FAILED);
     }
 
     @Test
-    @DisplayName("유저 정보 가져오기 실패 - 비밀번호가 틀림")
-    void getUserInfoPasswordIncorrect() {
+    @DisplayName("로그인 실패 - 비밀번호가 틀림")
+    void signInFailedPasswordIncorrect() {
         // given
         Set<AuthorityStatus> roles = new HashSet<>();
         roles.add(ROLE_MEMBER);
@@ -112,73 +189,20 @@ class UserServiceTest {
                 .updatedAt(LocalDateTime.now())
                 .build();
 
-        given(tokenAuthenticator.resolveTokenFromHeader(any())).willReturn("asdf1234");
         given(userRepository.findById(anyString())).willReturn(Optional.of(user));
 
         // when
         CustomException exception = assertThrows(
                 CustomException.class,
-                () -> userService.getUserInfo(any(), "qwer1233")
+                () -> authService.signIn(
+                        SignInDto.Request.builder()
+                                .id(user.getId())
+                                .password("qwer1233")
+                                .build())
         );
 
         // then
         assertEquals(exception.getErrorCode(), ErrorCode.PASSWORD_INCORRECT);
     }
 
-    @Test
-    @DisplayName("유저 정보 수정 성공")
-    void updateUserInfoSuccess() {
-        // given
-        Set<AuthorityStatus> roles = new HashSet<>();
-        roles.add(ROLE_MEMBER);
-
-        User user = User.builder()
-                .id("asdf1234")
-                .password("qwer1234")
-                .roles(roles)
-                .registeredAt(LocalDateTime.now())
-                .updatedAt(LocalDateTime.now())
-                .build();
-
-        String newPassword = "zxcv1122";
-
-        given(tokenAuthenticator.resolveTokenFromHeader(any())).willReturn("asdf1234");
-        given(userRepository.findById(anyString())).willReturn(Optional.of(user));
-
-        // when
-        UserDto userDto = userService.updateUserInfo(any(), UpdateUserInfo.builder()
-                .password("qwer1234")
-                .newPassword(newPassword)
-                .build());
-
-        // then
-        assertEquals(userDto.getPassword(), newPassword);
-    }
-
-    @Test
-    @DisplayName("유저 삭제 성공")
-    void deleteUserInfoSuccess() {
-        // given
-        Set<AuthorityStatus> roles = new HashSet<>();
-        roles.add(ROLE_MEMBER);
-
-        User user = User.builder()
-                .id("asdf1234")
-                .password("qwer1234")
-                .roles(roles)
-                .registeredAt(LocalDateTime.now())
-                .updatedAt(LocalDateTime.now())
-                .build();
-
-        given(tokenAuthenticator.resolveTokenFromHeader(any())).willReturn("asdf1234");
-        given(userRepository.findById(anyString())).willReturn(Optional.of(user));
-
-        ArgumentCaptor<User> captor = ArgumentCaptor.forClass(User.class);
-
-        // when
-        userService.deleteUserInfo(any(), "qwer1234");
-
-        // then
-        verify(userRepository, times(1)).delete(captor.capture());
-    }
 }
